Исправлены ошибки, при работе с int-ми полями.
git-svn-id: https://share.curs.ru/svn/grains/celestabasic/trunk/common.sys@1639 bf7a05a2-0d8f-8241-91a8-991a96a8b1b0

+++ score/common/filtertools/filter_cards_gen.py
@@ -28,7 +28,7 @@ def card_save(xformsdata, context, filter_id):
             context.getData()[u'card_save'] = set([])
         context.getData()[u'card_save'].add(filter_id)    
         temp_context = xformsdata["schema"]["filters"]['filter']
-        handle = {'float': lambda x: float(x) if x.isdigit() else '', 
+        handle = {'float': lambda x: int(x) if x.isdigit() else '', 
                   'text': unicode, 
                   'bool': unicode, 
                   'date': unicode    # Потом везде сделать нормальную обработку
